WIP upgrading logger to 1.5 and adding console appender as default for client scripts.
Note that for the moment I had to hand edit the path to
the aurelia-logging dependency in node_modules\aurelia-logging-console\dist\aurelia-logging-console.js
to be relative "./aurelia..." so it would be found when loaded in NS.

TODO: figure out why the autologger is logging to execution log even in client scripts. I think it is due to when that is initialized the runtime.executionContext is not set?