simplify unit testing configuration to work again for zach and others.
The 'moduledirectories' including 'netsuite-fasttrack-toolkit-ss2' was not going to work because that folder name isn't reliable (indeed, we named it just 'nft' by default' - so not sure how that go in there. Also, using the prior current directory isn't ideal either, so remove it and hoped newer versions will resolve paths through `node_modules` as indicated. It _seems_ to work - will need to try a clean clone to ensure.

governance.test.ts - remove an extra EC_Logger mock that we didn't appear to need for the test to pass anyway.

 also update moment to avoid vulnerability.